+++ app/src/main/java/com/pavel/a692group/presentation/login/AuthSingleFragmentActivity.java
@@ -1,10 +1,12 @@
-package com.pavel.a692group;
+package com.pavel.a692group.presentation.login;
 
 import android.os.Bundle;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.pavel.a692group.R;
 
 /**
  * Created by Pavel Mazhnik on 02.03.19.
