+++ app/src/main/res/layout/activity_info.xml
@@ -17,7 +17,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context="com.pavel.a692group.InfoActivity">
+        tools:context="com.pavel.a692group.presentation.message.InfoActivity">
 
 
         <Button
