+++ app/src/main/res/layout/activity_login.xml
@@ -17,7 +17,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context="com.pavel.a692group.LoginFragment">
+        tools:context="com.pavel.a692group.presentation.login.LoginFragment">
 
         <ScrollView
             android:id="@+id/login_form"
